+++ lib/fog/compute/kubevirt/requests/delete_vm.rb
@@ -2,13 +2,13 @@ module Fog
   module Compute
     class Kubevirt
       class Real
-        def delete_livevm(name, namespace)
+        def delete_vm(name, namespace)
           kubevirt_client.delete_virtual_machine(name, namespace)
         end
       end
 
       class Mock
-        def delete_livevm(name, namespace = nil)
+        def delete_vm(name)
         end
       end
     end