+++ lib/fog/compute/kubevirt/requests/list_vminstances.rb
@@ -1,0 +1,21 @@
+module Fog
+  module Compute
+    class Kubevirt
+      class Real
+        def list_vminstances(_filters = {})
+          vminstances = kubevirt_client.get_virtual_machine_instances(namespace: @namespace)
+          entities = vminstances.map do |kubevirt_obj|
+            Vminstance.parse object_to_hash(kubevirt_obj)
+          end
+          EntityCollection.new(vminstances.kind, vminstances.resourceVersion, entities)
+        end
+      end
+
+      class Mock
+        # TODO provide implementation
+        def list_vminstances(_filters = {})
+        end
+      end
+    end
+  end
+end